app/requirements.txt
requests
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
djangochannelsrestframework==1.1.0
django-celery-results
django-simple-history
drf-spectacular==0.24.1
django-celery-beat
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
phonenumbers==8.12.54
django-jet-reboot==1.3.3
celery==5.2.7
redis
django-storages==1.13.1
rabbitmq==0.2.0
channels-redis
daphne==4.0.0
pandas==1.4.4
rest-pandas
openpyxl==3.0.10
channels-redis
django-phonenumber-field